pyproject.toml
[project]
name = "inkwell"
version = "0.1.0"
description = "AI-powered email drafting application"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.12",
    "groq>=0.37.0",
    "python-dotenv>=1.2.1",
    "pydantic-settings>=2.12.0",
]

[project.scripts]
inkwell = "run:main"
